backend/requirements.txt
psycopg[binary]
python-dotenv
fastapi
uvicorn
SQLAlchemy
pwdlib[bcrypt]
pydantic[email]
PyJWT
slowapi
